Ascent: Add Edges game mode.
All numbers in this mode are given as arrows on the edge of the grid.
The arrow points to the row, column or diagonal where the number is located.

This mode is an implementation of 1to25 by Jeff Widderich. While it supports
grids of any size, it plays best on grids of 5x5.