Ascent: Change dragging behaviour.
Clicking a number, then clicking it again to drag a path now does the same as dragging immediately. Previously, this would reverse the direction when the user is expecting to drag the path forward. To reverse the direction, the user has to click a highlighted number without dragging. The legacy behaviour can still be enabled by setting ui->fast_turn to TRUE.